backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]>=0.24.0,<1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy>=2.0.36,<3.0.0
alembic==1.12.1
# Using SQLite for development - lightweight and no additional setup required
aiosqlite==0.19.0
# PostgreSQL driver (commented out by default)
# psycopg2-binary==2.9.9
# Cloud services
google-cloud-firestore==2.13.0
google-cloud-bigquery==3.14.1
# AI and ML
openai>=1.30.0,<2.0.0
# Data validation
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
email-validator>=2.1.0,<3.0.0
